Fix 2

+++ tasks/baranov_a_dijkstra_crs/mpi/src/ops_mpi.cpp
@@ -46,7 +46,7 @@ bool BaranovADijkstraCrsMPI::ValidateGraph(const GraphCRS &graph) {
   if (graph.source < 0 || static_cast<size_t>(graph.source) >= static_cast<size_t>(graph.vertices)) {
     return false;
   }
-  if (graph.row_ptr.empty() || graph.row_ptr.size() != graph.vertices + 1) {
+  if (graph.row_ptr.empty() || graph.row_ptr.size() != static_cast<size_t>(graph.vertices) + 1) {
     return false;
   }
